chg: Move Navbar to handle the account case to simplify it
