+++ app/page/home/index.tsx
@@ -1,7 +1,9 @@
+import { NetworkSelector } from "app/page/home/navbar"
 import { Wallet } from "ethers"
 import { useContext, useEffect, useState } from "react"
 import ArrowDown from "react:~assets/arrowDown.svg"
 import ArrowUp from "react:~assets/arrowUp.svg"
+import ChevronDown from "react:~assets/chevronDown.svg"
 import { Link } from "wouter"
 import { useHashLocation } from "wouter/use-hash-location"
 
@@ -31,11 +33,11 @@ export function Home() {
   const hasNoAccount = useAccounts().length === 0
   return (
     <>
-      <Navbar />
       {hasNoAccount ? (
         <AccountCreation />
       ) : (
         <>
+          <Navbar />
           <AccountInfo />
           <Divider />
           <AccountNavigation />
@@ -70,7 +72,35 @@ function AccountCreation() {
     initSimpleAccount()
   }, [network.id])
 
-  return <></>
+  return (
+    <div className="flex items-center justify-between mb-[16px] mt-[4px]">
+      <div>
+        <NullAccountSelector />
+      </div>
+      <div>
+        <NetworkSelector />
+      </div>
+    </div>
+  )
+}
+
+function NullAccountSelector() {
+  const [, navigate] = useHashLocation()
+  return (
+    <>
+      {/* Home page account selector button */}
+      <button
+        className="px-[20px] py-[14.3px] flex items-center rounded-full border-[1px] border-solid border-black"
+        onClick={() => navigate(Path.AccountList)}>
+        <div className="mr-[12px] flex flex-col items-start">
+          <div className="leading-[19.4px] text-[16px] text-[#000000] whitespace-nowrap">
+            No account
+          </div>
+        </div>
+        <ChevronDown className="w-[16px] h-[16px]" />
+      </button>
+    </>
+  )
 }
 
 export function AccountInfo() {
